update origination fee parameters from GammaPoolFactory

+++ contracts/test/strategies/base/TestRebalanceStrategy.sol
@@ -14,7 +14,7 @@ contract TestRebalanceStrategy is BorrowStrategy, RebalanceStrategy {
     event LoanCreated(address indexed caller, uint256 tokenId);
     event AmountsWithFees(uint256[] amounts);
     uint80 public borrowRate = 1e18;
-    uint24 public origFee = 0;
+    uint16 public origFee = 0;
     uint16 public protocolId;
     uint256 private _minBorrow = 1e3;
     uint256 private mCurrPrice = 1e18;
@@ -249,11 +249,11 @@ contract TestRebalanceStrategy is BorrowStrategy, RebalanceStrategy {
         CFMM_RESERVES = s.CFMM_RESERVES;
     }
 
-    function setOriginationFee(uint24 _origFee) external virtual {
+    function setOriginationFee(uint16 _origFee) external virtual {
         origFee = _origFee;
     }
 
-    function originationFee() internal override virtual view returns(uint24) {
+    function originationFee() internal override virtual view returns(uint16) {
         return origFee;
     }
